pushing layout template file
This is a layout for news article page, it has the full template of the
website, other pages will only add their own content representation in
the container.
The template has some issues concerning the sliding menu with some
browsers, some rendering issues specially when scrolling, also the menu
button trigger is currently the whole bar not the menu button.